cred/cred/requirements.txt
djangorestframework==3.11.1
uuid==1.30
requests==2.24.0